+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.32
+### is.codion.swing.common.ui
+- ComponentBuilder implementations, lambdas replaced with static classes to prevent memory leaks.
 ### is.codion.swing.framework.ui
 - EntitySearchField.selectorFactory() now Value based.
 - EntitySearchField.searchIndicator() now Value based.